fixme: other speed ups across other commands.
