+++ samcli/commands/local/cli_common/mode.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class ContainersInitializationMode(Enum):
+    EAGER = "EAGER"
+    LAZY = "LAZY"
+
+
+class ContainersMode(Enum):
+    WARM = "WARM"
+    COLD = "COLD"
