+++ samcli/lib/sync/sync_flow_factory.py
@@ -16,7 +16,6 @@ from samcli.lib.sync.flows.image_function_sync_flow import ImageFunctionSyncFlow
 from samcli.lib.sync.flows.rest_api_sync_flow import RestApiSyncFlow
 from samcli.lib.sync.flows.http_api_sync_flow import HttpApiSyncFlow
 from samcli.lib.sync.flows.stepfunctions_sync_flow import StepFunctionsSyncFlow
-from samcli.lib.utils.boto_utils import get_boto_resource_provider_with_config, get_boto_client_provider_with_config
 from samcli.lib.utils.cloudformation import get_resource_summaries
 from samcli.lib.utils.resources import (
     AWS_SERVERLESS_FUNCTION,
@@ -36,35 +35,6 @@ if TYPE_CHECKING:  # pragma: no cover
     from samcli.commands.build.build_context import BuildContext
 
 LOG = logging.getLogger(__name__)
-
-
-class SyncCodeResources:
-    """
-    A class that records the supported resource types that can perform sync --code
-    """
-
-    _accepted_resources = [
-        AWS_SERVERLESS_FUNCTION,
-        AWS_LAMBDA_FUNCTION,
-        AWS_SERVERLESS_LAYERVERSION,
-        AWS_LAMBDA_LAYERVERSION,
-        AWS_SERVERLESS_API,
-        AWS_APIGATEWAY_RESTAPI,
-        AWS_SERVERLESS_HTTPAPI,
-        AWS_APIGATEWAY_V2_API,
-        AWS_SERVERLESS_STATEMACHINE,
-        AWS_STEPFUNCTIONS_STATEMACHINE,
-    ]
-
-    @classmethod
-    def values(cls) -> List[str]:
-        """
-        A class getter to retrieve the accepted resource list
-
-        Returns: List[str]
-            The accepted resources list
-        """
-        return cls._accepted_resources
 
 
 class SyncFlowFactory(ResourceTypeBasedFactory[SyncFlow]):  # pylint: disable=E1136
@@ -103,6 +73,9 @@ class SyncFlowFactory(ResourceTypeBasedFactory[SyncFlow]):  # pylint: disable=E1
     def load_physical_id_mapping(self) -> None:
         """Load physical IDs of the stack resources from remote"""
         LOG.debug("Loading physical ID mapping")
+        from samcli.lib.utils.boto_utils import get_boto_resource_provider_with_config, \
+            get_boto_client_provider_with_config
+
         resource_provider = get_boto_resource_provider_with_config(
             region=self._deploy_context.region, profile=self._deploy_context.profile
         )
